chore: split housing navigator form into new components
